+++ beam/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log - @stonecrop/beam
 
+## 0.37.0
+
 ## 0.36.0
 
 ## 0.35.0