+++ nuxt_grafserv/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log - @stonecrop/nuxt-grafserv
 
+## 0.37.0
+
+### Patch Changes
+
+- @stonecrop/casl-middleware@0.37.0
+  - @stonecrop/graphql-middleware@0.37.0
+  - @stonecrop/rockfoil@0.37.0
+
 ## 0.36.0
 
 ### Patch Changes